Fix tests

+++ tests/geometry/AngleCmp.test.cpp
@@ -38,5 +38,5 @@ int main() {
     p.x = x; p.y = y;
   }
   sort(a.begin(), a.end(), angle_cmp);
-  for (P p : a) cout << p.x << ' ' << p.y << '\n';
+  for (P p : a) cout << (int)p.x << ' ' << (int)p.y << '\n';
 }
